backend/requirements.txt
# FastAPI and web server
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.6,<1.0.0

# Database
sqlalchemy[asyncio]>=2.0.25,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.1.0,<5.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Configuration
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# HTTP client (for OAuth)
httpx>=0.26.0,<1.0.0

# Stripe integration
stripe>=7.0.0,<8.0.0

# Redis (for sessions/caching)
redis>=5.0.0,<6.0.0

# Task queue
celery[redis]>=5.3.0,<6.0.0

# S3-compatible storage (DigitalOcean Spaces, AWS S3)
boto3>=1.34.0,<2.0.0

# Image processing
pillow>=12.1.0,<14.0.0
opencv-python-headless>=4.9.0,<5.0.0
numpy>=2.3.0,<3.0.0

# AI/ML
fal-client>=0.4.0,<1.0.0
requests>=2.31.0,<3.0.0

# Utilities
email-validator>=2.1.0,<3.0.0

# Error tracking
sentry-sdk[fastapi]>=1.40.0,<3.0.0

# Development and testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0
httpx>=0.26.0  # Also used for async test client
